fixed a bug of double boot

+++ bot/ai.js
@@ -7,13 +7,14 @@ const logger = new Logger("ai");
 
 client.on("messageCreate", async message => {
     if (message.author.bot) return;
-    if (message.content.includes(`<@${client.user.id}>`) && (message.content !== `<@${client.user.id}>`) || (message.reference && Model.chatData.has(message.reference.messageId))) {
+    if (message.content.includes(`<@${client.user.id}>`) && (message.content !== `<@${client.user.id}>`) || (message.reference && Model.chatData.has(message.reference.messageId) && message.mentions.repliedUser)) {
         try {
             if (!Model.current.ready) return await message.channel.send("```モデルは準備中です```")
             message.channel.sendTyping();
             let ids;
             if (message.reference) ids = Model.chatData.get(message.reference?.messageId);
             let image = message.attachments?.first();
+            message.content = message.content.split(`<@${client.user.id}>`).join(Model.current.name + "、");
             const res = await Model.current.ask(message, ids ,image);
             const successful = await message.channel.send("✓");
             await successful.delete();
